+++ src/config/constants.py
@@ -4,7 +4,7 @@ Application Constants
 
 APP_NAME = "Book Library Reconciler"
 
-VERSION = "2.0.0-alpha.6"
+VERSION = "2.0.0-alpha.7"
 
 # --------------------------------------------------
 # Display